Add custom icon
Note: Pixel-art needs to be upscaled before being converted into an .ico file, to not be blurry. I used LibreSprite to make and upscale this image, starting with 9x9 and upscaling it to its current resolution using the internal scaling tools. Upscaling tools not made specifically for pixel-art tends to apply a smoothing effect.